feat: re-org + generic orchestrator rewrite
